+++ src/thumbor_video_engine/filters/format.py
@@ -2,7 +2,6 @@ from thumbor.filters import BaseFilter, filter_method, PHASE_PRE_LOAD
 import thumbor.filters.format
 from thumbor.utils import logger
 
-from thumbor_video_engine.compat import filter_retval
 
 
 ALLOWED_FORMATS = list(thumbor.filters.format.ALLOWED_FORMATS) + [
@@ -19,7 +18,7 @@ class Filter(BaseFilter):
     phase = PHASE_PRE_LOAD
 
     @filter_method(BaseFilter.String)
-    def format(self, format):
+    async def format(self, format):
         logger.debug("Setting format to %s" % format)
         if format.lower() not in ALLOWED_FORMATS:
             logger.warning("Format not allowed: %s" % format.lower())
@@ -27,5 +26,3 @@ class Filter(BaseFilter):
         else:
             logger.debug("Format specified: %s" % format.lower())
             self.context.request.format = format.lower()
-
-        return filter_retval()